+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `Ctrl-r` to cycle repeat mode
 - Fix duplicate albums showing in artist discographies
+- Add `n/p` to skip to previous or next track
+
 
 ## [0.0.2] - 2019-09-17